[6316] Read table column widths
I noticed this wasn’t done yet, and it was a simple change.

Note the column widths (which are doubles) are now displayed in the
native for the tests: is that ok? I wonder about rounding errors. Would
it be better to use a `Rational` here instead (especially since I’m
creating them by division)?