Fix table field detection
Two problems:
- it wasn’t detecting complex table fields
- the caption style check also caught captions for other things, e.g.
  equations

Both fixed: I now require the caption style /and/ either the simple or
complex table field. This means the caption detection fails for captions
which don’t contain a field, as in an example doc I added as a test.
However, I think it’s better to be too conservative: a missed table
caption will still show up as a paragraph next to the table, whereas if
I incorrectly classify something else as a table caption it could cause
havoc by pairing it up with a table it’s not at all related to, or
dropping it entirely.