Add test with ordered list inside a table caption
Because the numbering is handled through the DState part of DocxContext,
reading the captions in the first pass and the rest of the document in a
second pass means that a continued numbered list within the caption
won’t be numbered correctly.

I tried to create a word document with such an example, but creating the
numbered list creates a new paragraph, which the caption-reading code
doesn’t interpret as part of the caption.

So: allowing block-level content within captions will take a bit more
work, probably condensing neighbouring paragraphs with the caption style
into a single caption (as long as only one contains a table field, I
suppose?). Once we do allow that, we’ll need to change how we handle the
captions, probably by seeking through the document or something similar.